Add polyfill.io bundle to js modules build
Adds an uncompressed bundle of polyfills for
JS language features used by this version of
DiffDOM. It includes comments to the top, added by
Polyfill.io, giving details of all polyfills
included.

The code being polyfilled is 3rd party so we are
having to use this process to decide which
features to polyfill:
1. use BabelJS to output the CoreJS polyfills
   needed for the DiffDOM code
2. request a bundle from polyfill.io for the
   features the CoreJS polyfills cover
3. do some manual checks of the features
   polyfill.io doesn't include polyfills for

The polyfill.io bundle does not include polyfills
for these features (which CoreJS did) because IE9
has versions of all of them similar enough for
most usecases:

- Array.prototype.concat
- Array.prototype.join
- Array.prototype.slice
- Array.prototype.splice
- Object.prototype.toString
- RegExp.prototype.exec
- RegExp.prototype.toString
- String.prototype.match
- String.prototype.replace
- String.prototype.split

(Based on data from
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference.)

The latest version of the following features have
significant differences to the ECMAScript 3 (ES3)
versions but those more modern aspects are not
used by the DiffDOM code so are not included:

- JSON.Stringify
- RegExp.prototype.constructor (concerning the
  RegExp instances returned)
- RegExp.prototype.sticky

The modern JSON Stringify handles Symbols and has
protections for use of lone surrogate characters.
DiffDOM doesn't use Symbols and we control the
HTML it operates on so using the IE9 version isn't
risky.

Modern RegExp instances include a wider API and
support extra regular expression features (ie.
named capture groups). The API used by DiffDOM is
limited to that supported by the IE9 version,
based on this MDN reference:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp#browser_compatibility

Notes:

CoreJS vs Polyfill.io

The only time CoreJS will not polyfill a feature
is when the target browser includes a version of
it that matches the latest ECMAScript spec'.

For example, IE9 includes a version of
String.prototype.match from ES3 so it would be
polyfilled whereas Chrome has had the latest one
since 51 so would not.

Polyfill.io will not polyfill a feature if the
target browser already supports a version unless
the polyfill includes what it sees as important
changes. For example, IE9 support
Array.prototype.sort but it isn't guaranteed to
produce consistent results. The latest version
does so the polyfill for it is included.

We use Polyfill.io to filter our list of CoreJS
polyfills down to a list that favours browser
support over correctness to the spec'.

The polyfill.io bundle

I built the polyfill.io bundle by sending
a request to their API, for a non-minified file:
- specifying all the features
- setting flags to `gated`
- with a IE9 user agent string User-Agent header

See their API docs:

https://polyfill.io/v3/api/#request

List of user agent strings:

https://developers.whatismybrowser.com/useragents/explore/software_name/internet-explorer/

Detecting language features using BabelJS

I used BabelJS's preset-env with `useBuiltIns` set
to `usage` to analyse the language features
DiffDOM uses. It's REPL can show how this works:

https://babeljs.io/repl#?browsers=ie%209&build=&builtIns=usage&corejs=3.6&spec=false&loose=false&code_lz=MYewdgziA2CmB00QHMAUAoABJg2gRgBoAmAgZgF14BLMYaAVwBNYJU8BKddgbnXVEgwESNFkwB5AEYArWMAAu8WGHkAnKi1QBvAOQAzECB0AuTDskBDVToC-nHkA&debug=false&forceAllTransforms=false&shippedProposals=false&circleciRepo=&evaluate=false&fileSize=false&timeTravel=false&sourceType=module&lineWrap=true&presets=env%2Cstage-2&prettier=false&targets=&version=7.16.7&externalPlugins=&assumptions=%7B%7Du

Also adds a .jshintignore file to exclude the
polyfills code from our linting.