Add diffDOM to build with BabelJS transpiling
The vendored version of DiffDOM is written in
ECMAScript 6 (ES6) but we need to ship code in
ECMAScript 3 to support the browsers in our
matrix.

This requires:
- the polyfilling of ES6 language features
- transpiling of language syntax from ES6 or ES5

An example of language features introduced by ES6
is `Array.from`.

An example of language syntax introduced by ES6 is
the `let` and `const` keywords.

Determining the polyfills needed is a complex task
so I've split it off into a separate commit.

Transpiling ES6 syntax just requires:
1. adding BabelJS to the part of the frontend
   build that DiffDOM sits in
2. enabling BabelJS's `preset-env`, set to the
   default configuration

...so we do that here.

Also includes a bump of the BabelJS libraries to
the latest version.