Change how file upload is enhanced by CSS/JS
The current file upload component works by having
3 labels:
1. the original one, visually hidden
2. one styled like a button
3. one to contain the name of any files selected

Note: label 3 hasn't been used since we started
submitting the form as soon as a file is selected:

https://github.com/alphagov/notifications-admin/commit/97a3bf922580f61c81f386f9665a3f05724f4ec4

The label styled like a button uses the fact that
clicks on labels activate the input they label.
This means clicking on it is the same as clicking
on the input: bringing up the OS file picker
dialog.

There are a few problems with this:
- the label styled like a button is still
  semantically a label, not a button
- labels can't be focused (we hack this by giving
  it the focus style when the input it labels is
  focused so it looks like it is)
- DAC raised an issue against us using multiple
  labels, saying some assistive technologies don't
  support that

The second problem is only a problem because
you're not actually focusing what it looks like
you are, so what is announced isn't quite the same
as the text you see (in the button).

These changes take a different approach, having
only one label and using JS to
- add a real button
- hide the file upload input and its label
- pass any event that activates the button as a
  click on the input

This has the right number of labels and because
the button is a real button it has the right
semantics so assistive tech' can recognise it.
Screen readers will announce it correctly and
voice recognition software will let you click it
using the button text. Using a real button also
means it behaves as expected with no need for ARIA
or JS.

It's a bit more reliant on JS than the last
approach but it is simpler and works more
consistently with assistive tech'.

This also adds a label to the button for errors if
present.

Buttons don't need labels by default so this adds
one, to contain the errors, only if errors exist
and shows it to match the existing visual design.

Includes a change to the structure of the JS tests
to improve how they read. Originally suggested in
this comment on the pull request for the first go
at this work:

https://github.com/alphagov/notifications-admin/pull/4268#discussion_r902810835