Change content seen when file is uploading
The file-upload's cancel button is a link, styled
as a button so needs:
- its semantics to match its style (via
  role=button)
- to work like a button (via GOVUK Frontend JS,
  that adds the missing behaviours)

When the upload button is replaced with the cancel
button, this change is only communicated visually
and focus is lost. so I've added some content to
explain that uploading has started and that you
can use the cancel button to stop it. I also move
focus to this content which:
- keeps focus in the same part of the page
- announces the explanation