Make cancel button work with space bar by using the GOVUK Frontend button component JS.
We still need to ensure users know:
1. selecting a file will also upload it
2. when the upload has started
3. that they can cancel the upload by pressing the
   cancel button

Having text prefixing the cancel button fixes 2.
and 3. in the aural experience (with screen
readers) but adds a shift in content that is
distracting to visual users.

I don't know how to communicate 1. and need input
from interaction and content design. But I don't
want to block the work on this branch being
merged so I've held off making any other changes.

TO DOs:

Communicate everything in the list at the top of
this commit.

We also need to give non-visual users more
information about what is happening when the page
reloads with the file in it. We should be adding
something to the page title, like we do with
validation, and giving the file some alt text, if
it is an image.