Upgrade xlrd to version 2
xlrd is a library for reading data and formatting information from
Excel files in the historical .xls format.

Version 2 of xlrd no longer supports anything other than .xls files.

We were using it to also support reading .xlsm files (old Excel files
with macro support).

We could keep using the old version of this dependency, but hopefully
this niche version of an ancient file format is obscure enough that
no-one is using it, and we can drop support, keeping our dependencies
up to date.