WIP: Cache get_user_template_folders
This gives a new average render time of a little above 3s, down
from a little above 9s.

I'm not aware of a handy Python cache that's instance-specific, so
this is a quick implementation of one from scratch.

Note that, as with uses of @cached_property, we rely on the fact
that all instances are refreshed on every request.