Work around inconsistent request context in tests
The failing test here[1] does two things:
1. makes a request to /sign-out
2. calls the index route, without actually making a request

This means that when the `login_manager.unauthorized_handler`[2] looks
at Flask’s `request` object it gets the request context from 1., not 2.,
because 2. isn’t actually a request. The means it sets the value of the
`next` parameter to that of the request, not of the index route.

Basically at some point Flask has changed and decided that 2. isn’t a
proper request, so won’t set new request context.

This isn’t a realistic test because nothing would call the index
function directly, it would always be as part of a request to that page.

But to make the minimal change to fix the breaking tests this commit
makes the check a bit more general, i.e. that the redirect is to the
sign in page with any `next` parameter, not a specific `next` parameter.

1. https://github.com/alphagov/notifications-admin/blob/9111a7fc86b4ed2f88f9f913c0ea5bfe98f20ead/tests/app/utils/test_user.py#L130-L138
2. https://github.com/alphagov/notifications-admin/blob/9111a7fc86b4ed2f88f9f913c0ea5bfe98f20ead/app/main/views/sign_in.py#L86