Refactor listing user templates into new class
Previously I found it hard to understand TemplateList due to the
additional complexity of the "user" flag, which would switch the
class to work on a refined set of templates and folders.

This moves all the user-specific behaviour into a subclass. It's
still complex, but I think the new approach is simpler overall.
Among all the class renamings, the code to focus on is:

- In "get_template_folders" we're using the inheritance to break
apart the conditional on "self.user", with each branch moving to
new "all_template_folders" properties in the base and subclasses.

- In "get_templates" we're moving the conditional on "self.user"
directly to the new subclass. We'll do more work on this later so
it follows a similar (and therefore, simpler) pattern to folders.