Document +882 numbers are not supported
These are "countryless" phone numbers used by e.g. satellite phone
providers [^1]. We decided not to support them for now because:

- The use case is low: only one service is asking for this prefix.
- MMG will charge at 4x for them but apparently they cost more.

We can't put this "negative" data in the usual place [^2] because
this would _enable_ sending via this number. Until we have more
cases like this it's easiest to just tack on this fake info.

[^1]: https://en.wikipedia.org/wiki/International_Networks_(country_code)
[^2]: https://github.com/alphagov/notifications-utils/blob/ca2506c6a626d25f76a25c4a94b7288e1c3aa2b5/notifications_utils/recipients.py#L569