Pin govuk-frontend-jinja to version published by CCS
The version of govuk-frontend-jinja we are using was developed in GDS
and is now owned by Crown Commercial Service.

It doesn’t seem to be maintained any more.

There is a different package with the same name which is owned by the
Land Registry and is maintained. It is published on PyPi, so PyUp tries
to upgrade to it. We can’t automatically upgrade to it because it’s not
a like-for-like replacement for the Crown Commercial Service package.

This commit:
- clarifies that it’s the Crown Commercial Service package we are using
  (the alphagov URL redirects to this)
- tells PyUp to stop trying to upgrade the package