Set google analytics cookies as 'Secure'
We are currently seeing warnings in the javascript console when
using firefox of the following:

```
Cookie “_ga” will be soon rejected because it has the “SameSite” attribute set to “None” or an invalid value, without the “secure” attribute. To know more about the “SameSite“ attribute, read https://developer.mozilla.org/docs/Web/HTTP/Headers/Set-Cookie/SameSite
```

and the same for the `_gid` cookie.

We can fix these errors by setting the `Secure` attribute
on the two google analytics cookies.

Note, our other two cookies (`notify_admin_session` and `cookies_policy`)
already have the `Secure` attribute.

Note, we are only seeing these errors in Firefox and not in Chrome,
Edge or Safari because Firefox has different levels of compliance
for the latest web standards:
https://hacks.mozilla.org/2020/08/changes-to-samesite-cookie-behavior/

We don't set the SameSite value on our cookies. Firefox,
sets the default value as `None`. Firefox
will output the deprecation warnings that we are seeing as therefore
`SameSite = None` and `Secure` is not set. This is not an issue in
Chrome or Edge because their default value for `SameSite` is `Lax`
and so we do not have the combination of `SameSite = None` and
`Secure` not set.
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie/SameSite#browser_compatibility

We don't think this change will make any fundamental differences
to our security but it will at least remove the deprecation warnings.
It theoretically will make things more secure if users were visiting
our site using HTTP but this shouldn't really be a thing.