Add docstrings to TemplateList iterator classes
In response to [^1].

I haven't added docstrings to the TemplateListItem hierarchy of
classes as I think these are self-explanatory:

- They are at the bottom of the file, so it's likely someone will
have the context they need from earlier comments.

- The  TemplateListService class is referenced directly in the
comment of the ServiceTemplateList class that uses it.

We are also going to look at refactoring of the "...Service..."
classes into a presenter pattern, so I don't think it's worth the
time making the comments more comprehensive now.

[^1]: https://github.com/alphagov/notifications-admin/pull/4258#discussion_r890065674